[Azure.Core] Extend Test Environment with Storage Endpoint (#13812)
The focus of these changes is to expose the Storage Endpoint Suffix for
the active Azure environment to the execution environment for tests. This
value differs between cloud environments and there is no convenient way to
query it from the data plane.

This is needed to allow for building a full connection string for a storage account
within within the scope of a test execution.