[Event Hubs Client] Better Handling of Connection Closing Scenario
The focus of these changes is to better handle the scenario where the
connection is closing when a link is being created or authorization is
being sent in the background.  Previously, there was a bug that prevented the
ObjectDisposedException from being wrapped in a more developer-friendly
exception, leading to a less clear message and a longer path for retries.

This new path will ensure that the connection state is queried when sending
the authorization token and that the result is properly awaited so that the
exception that can occur for the race condition is appropriately handled.