Adding new path-based rule for the Compute API (#13730)
This is a new yml configuration file for the OpenAPI's SLA bot. Once the bot is enabled on the repo, this rule in the file specifies a configuration helpful to the cplat (Compute Platform) sdk team. This configuration rule mainly specifies a path, here the Compute Management library's sdk path, and an email that will be sent by the bot whenever a PR is opened and the PR changes are in the Compute Management library's path.