[Event Hubs Client] Idempotent Producer Client
The focus of these changes is implementing the idempotent publising feature
infrastructure into the `EventHubProducerClient` and associated types,
refactoring the current structure to support idempotent and non-idempotent
publishing as unqiue code paths.

Not included in these changes is the addition of the `ReadPartitionPublishingProperties`
member and live tests; those will be covered in dedicated work streams.