[Event Hubs Client] Prefetch Size In Bytes Option
The focus of these changes is to introduce an option for the various event
consumers allowing the prefetch cache to be filled based on a size-based
heuristic rather than a count of events.  This feature is considered a
special case, helpful in scenarios where the size of events being read is
not able to be known or predicted upfront and limiting resource use is
valued over consistent and predictable performance.