[Event Hubs Client] Integrate v5.3.0-beta.1 Release Branch (#15162) (#15181)
The focus of these changes is to integrate the release branch for Event Hubs
v5.3.0-beta.1.

The release introduces an option for the various event consumers allowing
the prefetch cache to be filled based on a size-based heuristic rather than
a count of events.  This feature is considered a special case, helpful in
scenarios where the size of events being read is not able to be known or
predicted upfront and limiting resource use is valued over consistent and
predictable performance.