[Event Hubs Client] Idempotent Producer Service Integration (#15332)
The focus of these changes is to complete the initial round of end-to-end
integration with the Event Hubs service.  Functionality has been unit tested
and confirmed to not cause regressions in the non-idempotent Live tests.

Live tests for the idempotent feature are not included; they will follow as
a dedicated unit of work pending local tests.  Until service-support for the
feature is rolled out to the public cloud, Live testing as part of nightly
runs will not be possible.