[Event Hubs Client] Idempotent Publishing Live Tests
The focus of these changes is to introduce the Live tests for the
idempotent publishing feature, fix an error in AMQP metadata names, and
address some delayed feedback from the previous set of changes.

Note that the Live tests are currently disabled as the feature is not yet
available across the public cloud and requires a special private resource
to test against.