Add delay for replication for live Search test runs
I un-ignored a test that should be fixed after the changes to Azure.Core.TestFramework @pakrym and I worked on.

The extra delay was recommended by @brjohnstmsft to give the replicas time to replicate. A recent live run also resulted in an error in AutoFlushInterval_PartialBatch, and I believe the delay should help there as well. I believe a replica may still be getting doc updates (insertions) when a delete request is replicated.