Allow use of refreshInterval for KV Secret Provider
In the migration from the previous KeyVaultSecretProvider repo, the access
restriction on the KVSPO object became internal. This means that there's a
lack of any meaning to the refreshInterval-handling code in the package as
it will always be null.

This change re-exposes the options object so that users can take advantage
of the functionality to automatically refresh secret values.