Mitigate breaking build using SaveDebugRecordingsOnFailure (#16599)
Tested with debug and release builds with and without SYSTEM_TEAMPROJECTID defined (used in a few projects for detecting AzDO). Doing this in the setter works when set in either a fixture's constructor or in a test method, while doing so in SetUp is too early, and in TearDown doesn't propagate.