[Azure.Core] Remove TestFramework Target Duplicate
The focus of these changes is to remove a redundant specification of
the `net461` target; one is implicity brought in as part of the
`$(RequiredTargetFrameworks)` variable and another was hard-coded into
the list.

When using Visual Studio, this was causing a warning in the Package Manager
output stating:

    Error occurred while restoring NuGet packages: Invalid restore input.
    The original target frameworks value must match the aliases. Original
    target frameworks: netcoreapp2.1;net461;net461;net47

While this wasn't impactful, removing the duplicate framework resolves it.