[Event Hubs Client] Test Timeout Tweak
The focus of these changes is to bump up the timeout applied to the
Event Hubs live test runs.  This is intended to help provide a benchmark
for how long the run is taking, as it has recently begun to exceed the
timeout for reasons under investigation. (and which are not observed locally)

The current suspicion is that the use of canary and/or throttling or
contention in the management plane are contributing.  Once a baseline
run time is understood, we'll experiment with the level of parallelism.

Also riding along are a few small documentation changes that were
accidentally not included with the last set of changes.