[Event Hubs Client] Live Test Troubleshooting Tweaks
The focus of these changes is to reduce parallelism in the test run to
a single suite at a time and reduce the maximum run time.  This is intended
to allow for farming the test run logs to track which tests have executed
before the timeout occurs, to try and reduce the number of failure candidates.