[Event Hubs Client] Test Prefetch Adjustments
The focus of these changes is to tweak configuration of the Event Hubs
clients in scenarios where the behavior of multiple readers acting on the
same partition are being validated.  These are susceptible to a race condition
where the prefetch queue size was large enough that all expected events
were held and the expected service calls did not take place, causing
intermittent test failures.