Ensure that type is populated when deserializing into a CloudEvent (#17468)
During UX studies, users were easily confused if they attempted to deserialize an event grid event into a CloudEvent as no exception was thrown. We intentionally removed the validation as part of addressing https://github.com/Azure/azure-sdk-for-net/issues/16012, but this may have gone too far in the other direction. A good compromise seems to leave in the validation that the service uses which is to ensure that Type is populated for CloudEvents, while allowing Source to be null.