[Event Hubs] Exclude Legacy Packages from Pipelines
The focus of these changes is to remove the legacy packages from the
build pipeline definitions, as they're now covered under a dedicated set
of pipeline definitions.

This is intended to allow the T1 and T2 Event Hubs packages to be treated
as separate units of work, as they will continue to evolve independently
and are maintained by different teams.