[EngSys] Allow Project Exclusions (#17658)
Extending the build pipeline to allow project exclusions to be specified
within the serivce directory.  This is intended to enable splitting the
pipeline into multiple definitions for service directories that contain
both T1 and T2 packages, particularly in the case where those are maintained
by different teams.

Also included in this set of changes is a configuration set for the legacy
Event Hubs packages.