[Service Bus Client] Pipeline Split
The focus of these changes is to split the Service Bus pipeline into multiple
units, one dedicated to the current generation client and another dedicated to
the legacy client.

This will ensure that build errors and test failures are scoped to their associated
package and do not cause cascading failures.