[Event Hubs Client] Internalize Idempotent Publishing
The focus of these changes is to internalize types and members related
to the idemtpotent publishing feature, as this will not be included in the
next Event Hubs release.

A correction was also made to the initial value of the last published
sequence number in the `PartitionPublishingProperties` to align with
service semantics.