[Messaging Clients] Test Resource RBAC Grants
The focus of these changes is to try and eliminate the potential for
duplicate role assignments during retries of template deployment.

The previous approach attempted to randomize the name sufficiently to avoid
collisions during retries; however, this increased failures because the
same grant cannot exist in the same scope for the same principal - regardless
of name.

This revision attempts to create a deterministic name based on the resource
and principal that the grant is associated with, and also to avoid performing
the assignment unless the key resources of the template were successful.