Ignore sample tests (#18292)
The sample tests weren't intended to be run in the CI, but were instead useful to demonstrate how you might perform the different methods of authentication. That's actually done in the README now, so i'm just going to remove it since I can't make the build happy with keeping them around.