[Event Hubs Client] Move Blob Checkpoint Store to Shared (#18395)
The focus of these changes is to complete migration of the Blobs Checkpoint
Store from the `Azure.Messaging.EventHubs.Processor` project to the
`Azure.Messaging.EventHubs.Shared` project.  To facilitate sharing the
checkpoint store between the processor and the Functions extensions, the
implementation for the store was moved into the shared project, though its
tests remained in the processor.  In order to follow the pattern of keeping
locality between shared code items and their tests, the checkpoint store
tests, both unit and live, have been migrated to the shared testing project.

In support of this, some of the live testing infrastructure specific to
managing storage resources has also been migrated into the shared project
as part of the `BlobTesting` category of shared items.

Some refactoring and reformatting of the storge tests has also been performed
to improve consistency with other Event Hubs code, trim dead areas, and
improve readability.