[EventHubs] Add EventProcessor<TPartition> Sample
This sample complements a soon to be published blog post on using
`EventProcessor<TPartition>` to build a custom event processor.

Because of how it is structured, the blog post is able to go in a bit
more depth over the semantics of each method from
`EventProcessor<TPartiton>` which is being implemented, but I've added
some light comments in the code to help readers which are viewing this
in isolation. We plan to link to this sample from the blog post, and
when the blog is published, I will add a backlink in the sample to the
post.