[Event Hubs Client] Storage Implementation Feedback
The focus of these changes is to consider feedback provided by an internal
review of the blob storage manager by one of the developers responsible for
the Azure Blob Storage client library.  Many of the larger and more critical
items have already been used for adjustments; these tweaks address the majority
of the remaining minor items.