[Event Hubs Client] Migration Guide Updates
The focus of these changes is to overhaul the format of the migration
guide and refine its content.  All of the guide code samples are now
driven from snippets, including those from the legacy library.

Due to code snippet regeneration and a formatting pass, some other minor
orthogonal changes are also included.