[Core.Experimental] Spike to add `JsonData`
As part of the low level client effort, we wanted to experiment with a
slightly different API design for our type which represents mutable
JSON Data. This spike adds an initial implementation. This is very
similar to `DynamicJson` but I decided to create another type because
I did not want to impact current uses of `DynamicJson` and we expect
that the shape of these APIs will continue to evolve as we do our user
studies.

There are still some untested methods but we have pretty good coverage
(and we'll push to improve it as we lock the design).

An initial pass at API documentation was also done, as we want to have
some level of documentation for users during our user study. A more in
depth pass will be done as we lock the design.