Update Sample5_SignVerify.md (#18590)
* Update Sample5_SignVerify.md

Renamed variabled `rsaKey` and `ecKey` to `...KeyOptions`, to better illustrate that the object does not hold the key but the config options that tell the `keyClient` how to create the new key.

* Rename variables to better illustrate the actual referenced objects

Rename variable `rsaKey` to `rsaKeyOptions` since that object does not hold the key but the options which tell the keyClient how to create the new RSA key. Rename `cloudRsaKey` to `rsaKey` since that variable contains a reference to a key object. Both changes analogous for EC keys.

* Update Sample5_SignVerify.md

Co-authored-by: Heath Stewart <heaths@microsoft.com>