[Smoke Tests] Refactor Test Resources ARM (#18810)
The focus of these changes is to refactor the test resources ARM template
to shift to using the namespace-level connection strings for the Event Hubs
resource to allow for a new sample to be added without the need to create
additional namespaces or auth rules.  This pattern is also more closely
aligned with the majority of customer scenarios in real-world use.