[Event Hubs Client] Fix for AMQP Race Condition
The focus of these changes is to address a race condition where the Event
Hubs service closes an AMQP connection or link after its state was validated
and an operation is being prepared.  When the AMQP library attempts to
initiate the operation, it fails returning an IllegalState condition.

In this scenario, the Event Hubs client library should consider the failure
to be transient and allow retries.  Our retry logic will ensure that the
client itself was not closed/disposed before allowing the retry attempt.

Because this exception path is also monitored as a benign race condition
when managing pooled transport producers, that code path has also been
validated to ensure no behavioral changes.  There were some minor formatting
tweaks in this area to better align with surrounding code.