[Low Level Client] Spike of a TextAnalytics client (#18746)
* [Core.Experimental] Spike to add `JsonData`

As part of the low level client effort, we wanted to experiment with a
slightly different API design for our type which represents mutable
JSON Data. This spike adds an initial implementation. This is very
similar to `DynamicJson` but I decided to create another type because
I did not want to impact current uses of `DynamicJson` and we expect
that the shape of these APIs will continue to evolve as we do our user
studies.

There are still some untested methods but we have pretty good coverage
(and we'll push to improve it as we lock the design).

An initial pass at API documentation was also done, as we want to have
some level of documentation for users during our user study. A more in
depth pass will be done as we lock the design.

* [Core.Experimental] Use `JsonData` in `DynamicRequest`

* [Low Level Client] Spike TextAnalytics prototype

This code was hand authored, but with an eye towards generating it
automatically. The API documentation is pulled verbatim from the REST
API documentation, as that's what we expect we would be able to do in
an autogenerated fashion.

* [Core.Experimental] Use dynamic instead of JsonData in Request/Response

* [Core.Experimental] Expose as both JsonData and dynamic

* [Low Level Client] Include README.md template

* [Core.Experimental] Respond to Krzysztof's Feedback

* [Core.Experimental] Remove DynamicJson in favor of JsonData

* [Low Level Client] Include README.md template

* [Core.Experimental] Update API Listing

* [Low Level Client] Implement User Study tasks

Co-authored-by: Chris Hamons <chris.hamons@microsoft.com>