[Core.Experimental] Tweak ToString() and add == overloads
Today the `ToString` behavior of `JsonData` is "return a string which
is the JSON encoding of the underlying object".  This is reasonable
but the implication is that `(new JsonData("foo")).ToString()` is
`"foo"` (note the quotes) instead of just foo.

This changes the behavior such that if the JsonData is an object or array
we JSON serialize it, if it's null we return "<null>" and otherwise we
just call .ToString() on the underlying value.

The demo samples have been cleaned up as part of this change.