[Event Hubs Client] Performance Tests (#18960)
The focus of these changes is to introduce a set of performance tests for
the Event Hubs client library, covering key scenarios for publishing and
reading events.

Notable omissions to the perforance scenarios are coverage for the
`EventProcessorClient` and `EventProcessor<TPartition>`.  To effectively test,
these require patterns not currently supported in the performance Testing
framework and will be addressed at a later time.