[Messaging Clients] Maximum Authorization Refresh
The focus of these changes is to apply a maximum duration to the time
allowed before authorization is refreshed with the Event Hubs or Service
Bus service.

This fixes an issue where some long-lived SAS tokens were exceeding the
maximum interval allowed by the refresh timer (49 days, 17 hours, and change),
causing exceptions in the messaging clients.