[Core.Experimental] Add JsonConverter for `JsonData`
The default serialization strategy for a CLR object is not what we
want for `JsonData`, since it works by serializaing all the public
properties of the type, and only a subset of the public properties are
valid for a given `JsonData` (for example, when it represents a
string, the `Properties` property throws `InvalidOperationException`
because `Properties` is only valid when you have a `JsonData`
representing an object.

One place this can really bite you is if you do something like:

```csharp
JsonData array = new JsonData(new [] { "hello", "world" });
JsonData fromPoco = new JsonData(new {
  property = array
});
```

When constructing a JsonData from a POCO, we first serialize it to
JSON and then use that JSON to initialize our `JsonData`, but without
a special converter, trying to serialize the value of the `property`
property will throw an InvalidOperationException.

Provide a converter that lets us control the serialization and
deserialization of `JsonData` so we can just emit or parse the
underlying JSON.