[Azure.Core] AzureNamedKeyCredential (#19569)
When using a shared key for authorization, some services require that
the name of the key be used as a component of the algorithm to form the
authorization token.  For these cases, `AzureKeyCredential` and
`AzureSasCredential` aren't adequate as they track only a single value.
The `AzureNamedKeyCredential` is intended to address this gap.