[TextAnalytics] Remove TA LLC Prototype
When we were doing out initial prototype of low level client, we picked TA as an example service we would use when building and validating our prototype because we knew the service and it was well suited for being a test bed for our efforts.

We are now focused on taking the stuff we learned from this experiment and using it for some other services, but this code doesn't need to stick around anymore, so we can remove it to reduce confusion.