[Purview] Generate Catalog and Scanning clients
This change adds two new dataplane packages,
Azure.Data.Purview.Catalog and Azure.Data.Purview.Scanning. The
clients are fully generated using our new "Low Level Client"
generator.

There's some more work to be done at the swagger level for both of
these services (for example, we may want to change the operation
groupings and we will need to rename some of the operation groups) but
that's work we'll want to try to do in the swagger itself so all
languages can get the improvements.

I did a small transform to ensure that the `endpoint` parameter for
these classes was of type Uri so it has the same shape as our
traditional high level clients.

We'd like to land these clients now (even though there will be some
possibly disruptive renamings over the next few weeks) so we can have
something to give to our partner teams that they can use to start
working on samples and other supporting collateral.