[Messaging] AMQP Exception Special Case Handling (#20440)
The focus of these changes is to follow advice offered by the maintainers
of the `Microsoft.Azure.Amqp` library to consider any `InvalidOperationException`
or derrivitive that occurs when opening an AMQP connection, link, or session
to be transient and allow retries.

This is needed because the AMQP library state becomes invalidated when
certain network conditions occur, such as the TCP socket being reset.  This
results in the AMQP library throwing when a connection or link is requested.
Rather than surfacing exceptions for the caller's context, an `InvalidOperationException`
or `ObjectDisposedException` is thrown to represent the internal state.