[Event Hubs Client] Detect Closed Shared Connection
The focus of these changes is to guard against a shared connection being
closed when performing service operations.  Previously, the consumer and
producer clients were not aware and would trigger exception behavior in
the AMQP library, the type of which could vary depending on its internal
state.

This scenario will now trigger a consistent `EventHubsException` that is
not retriable and offers an error message stating that the connection was
closed.