Remove filtering based on ci.yml
Given service bus is not in ci.yml but in ci.data.yml
the filtering currently excludes it from the project properties.

This commit is a temporary fix to unblock a current release and
shouldn't be merged into master. A more tested fix will be needed
from our eng/common workflow for the long-term fix.