[Event Hubs Processor] Stop attempting to steal inappropriately
The focus of these changes is to fix a bug in which load balancing
incorrectly identifies a need to attempt to steal a partition (even from
itself) and then fails to find an eligible candidate partition after running
unnecessary logic.  This manifests only as a log message for host applications,
as partitions were never actually stolen.