[Event Hubs Processor] Validate at Startup (#20861)
The focus of these changes is to introduce validation for configuration
and permissions errors that would prevent an event processor from being
able to recover without intervention.  For example, an incorrect connection
string or the inability to write to the storage container would be detected.

Validation is performed in an unobtrusive way intended to avoid delaying
the processor's ability to begin taking ownership of partitions while ensuring
that host applications are made aware of fatal problems within the scope of
the Start call, rather than having to rely on the error handler.

Also included in these changes is retiring of the ListCheckpointsAsync
implementation, which was deprecated in favor of GetCheckpointAsync previously.