[Event Hubs Client] Test Configuration Adjustments (#20888)
The focus of these changes is to adjust the clouds used for nightly and
manual test runs, reverting to the Public and Canary clouds.  Weekly runs
will include the UsGov and China sovereign clouds as well.   After discussion,
this aligns with the configuration that the engineering team intended.

Due to the China cloud performing slowly during runs, we've seen some tests
start to exceed the previous execution time cap; this cap has been increased
to reduce flakiness in weekly runs.  I'll continue to monitor and adjust as we
learn more about test stability and performance in the sovereign clouds.