[Event Hubs Client] Processor Stop Aborts Links
The focus of these changes is to detect that an event processor is attempting
to stop and force-close the AMQP links being used to read events.  This
should help to avoid pauses when stopping due to an active AMQP operation
in the transport consumer being blocked because no events were available.