[Core] Don't throw from `Response.Content` if memory stream is private
Relax our constraint that the underlying buffer of the MemoryStream
backing a buffered response be publicly visible. There are cases today
where `RequestBodyPolicy` will not allocate a new MemoryStream. One
example is if the response stream is already seekable (as is the case
in our PlaybackTransport) nd we've observed that there are likely
cases where `HttpClient` itself may use a MemoryStream as the response
stream without allowing the underling buffer to be exposed.

In cases where we can not extract the underlying body, just make a
copy of it. Since we know the underlying stream is a memory stream, we
don't need to worry about hidden IO.

Fixes #21048