[Event Hubs Client] Capture Unobserved AMQP Link Close Errors (#21426)
The AMQP transport library used by Event Hubs manages aspects of the transport
in the background, notably the flow of data into the prefetch queue.  In doing
so it controls attempts to read data, and detecting connection/link failures to
provide resiliency for callers.

Access to the AMQP link is managed through the `FaultTolerantAmqpObject<T>` which
uses a `GetOrCreate` pattern to handle faults and provide a resilient experience
for callers.  When a link was discarded due to a fault, it was not observable to
clients and a new link would be created on the next `GetOrCreate` call.

In the majority of scenarios, this behavior is desirable.  However, in some
scenarios - such as a partition being stolen by another event consumer asserting
exclusive access, the failure must be surfaced to callers so that the correct
remidiation can be taken.

The focus of these changes is to ensure that errors causing link termination
are surfaced via logging and that a stolen partition is treated as a special
case and is always surfaced to callers.

Because this behavior was a direct impact to the load balancing aspects of
the event processor, testing exposed some additional load balancing gaps that
were also addressed.  Lost partitions will now be detected more reliably and
the processor will proactively relinquish ownership and allow load balancing
to make recovery decisions.