[Event Hubs] Expect Closed Links When Stopping
The focus of these changes is to special-case recognition of the
`ObjectDisposedException` during link creation when an event processor is
stopping and treat it as an expected scenario.  This exception can occur
due to a race condition of the closing the connection scope while creation
of a link is already in-flight.