Remove readonly from Shared in HttpClientTransport
This is a workaround pending a proper design.
Some fault injection scenarios require changing the default
HttpClientTransport which is held in HttpClientTransport.Shared. We set
this to a new value using reflection. This work for the .NET Framework
runtime, but does not work in .NET Core. To get this to work in .NET
Core, the readonly modifer needs to be removed from the Shared field.