[Event Hubs] Incorporate AmqpAnnotatedMessage into EventData
The focus of these changes is to introduce the `AmqpAnnotatedMessage`
abstraction as the basis for `EventData`, allowing access to the full set
of AMQP message data.  This enables scenarios such as interaction with
other message brokers that make use of the protocol in ways that are not otherwise
exposed within the Event Hubs client.

This set of changes will be followed by refactoring of the `AmqpMessageConverter`
to work with the full AMQP message rather than the known `EventData` subset.