[Azure.Messaging.ServiceBus] Fix AmqpMessageBodyTests
The `DateTime` and `DateTimeOffset` test values in `AmqpMessageBodyTests.s_amqpValues` were created with `DateTimeOffset.Parse`, which would fail with the following exception when run on a system with a non-US culture:
```
System.TypeInitializationException : The type initializer for 'Azure.Core.Amqp.Tests.AmqpMessageBodyTests' threw an exception.
  ----> System.FormatException : String was not recognized as a valid DateTime.
   at Azure.Core.Amqp.Tests.AmqpMessageBodyTests..ctor()
--FormatException
   at System.DateTimeParse.Parse(String s, DateTimeFormatInfo dtfi, DateTimeStyles styles, TimeSpan& offset)
   at System.DateTimeOffset.Parse(String input)
   at Azure.Core.Amqp.Tests.AmqpMessageBodyTests..cctor()
```

By using `new DateTimeOffset(2021, 3, 24, 0, 0, 0, TimeSpan.Zero)` instead of `DateTimeOffset.Parse("3/24/21")` we ensure that the tests pass on any system regardless of its configured culture.