[Event Hubs] Client Identifier Option
The focus of these changes is to add an option for setting an Identifier
for each Event Hubs client type.  The identifier is informational and is
associated with the AMQP links used, allowing the service to provide additional
context in error messages and the SDK logs to provide an additional point
of correlation.