[Event Hubs] Processor Event Dispatch Cancellation (#22695)
The focus of these changes is to fix the loop responsible for dispatching
events to the handler for processing so that it respects cancellation when
the processor signals the associated token.  Currently, the loop will dispatch
all events in the batch to the handler, potentially causing duplicate processing
and checkpoint interleaving.