[Smoke Tests] Mitigate Failures
The focus of these changes is to mitigate the current failures for the
smoke tests.   To do so, references for packages have been updated to match
the latest August milestone deployments.  The `canary` Azure region has been
temporarily disabled due to test resource deployment issues that require a new
service version to fix.