build(sdk): enlarge memory of nodejs vm
`network` has 50 swagger files which will cause out-of-memory error in autorest.
Accoring to [autorest faq], we can enlarge the memory through environment variable.
Adding `memory` in autorest config doesn't work, because either release pipeline is not using the latest version of autorest, or the change is not in latest version yet.