[Event Hubs] Experimental Package
The focus of these changes is to add the `Azure.Messaging.EventHubs.Experimental`
pacakge for evaluating and testing types that may (or may not) eventually be promoted
to the Event Hubs client libraries.

In the initial version, a stand-alone Idempotent Producer has been included for
internal partner use and testing, built off the internalized API surface in the
core package.

The internal types in the core package that needed to be shadowed fully for use
have been renamed to avoid collisions in the namespace.