[Event Hubs] Buffered Producer Publishing Management
The focus of these changes is to implement the background tasks responsible for
managing the producer health and publishing batches of enqueued events.

Also riding along is an adjustment to the build props for the SDK directory,
removing the need to track individually per-package and locking the language
version to work-around a bug in the RC-1 compiler for .NET 6.

Pending work:
- Flush functionality
- Clear functionality
- Automatic partition assignment
- Partition Key hashing