[Event Hubs] Doc and Visibility Updates
The focus of these changes is to correct an error in the documentation
for the `EventProcessor<T>` and allow the `ToString` method for the
`EventPosition` type to appear in Intellisense.  In the latter case, `ToString`
is overridden to provide insight into what position in the event stream is
represented and has value to customers.