Fix live Key Vault tests (#24593)
* Fix live Key Vault tests

The services for the attestation mock service container were
conditioned, which required the `output` parameter to be conditioned and
it wasn't. Initially testing locally with enableHsm=true so didn't catch
it.

* Ignore tests that unnecessary require Attestation

Secure Key Release isn't yet supported on Key Vault, so ignore any tests that try to get this now-optional variable.