[Event Hubs] Buffered Producer Live Test and Fixes (#25147)
* [Event Hubs] Buffered Producer Live Test and Fixes

The focus of these changes is to implement the live tests for end-to-end
scenarios. Where the tests exposed flaws, they were corrected and a general
polish pass was done to the implementation.

This set of changes also makes the producer and associated support types
public in anticipation of the pending beta release. Also riding along are
some minor formatting changes.