[Azure.Core] Logging Sample Enhancement (#25389)
# Summary

The focus of these changes is to enhance the logging section of the diagnostics sample to include details about the lifetime of the `AzureEventSourceListener`.  We received some customer feedback that it was unclear where they needed to create the listener and that it had to have the appropriate scope to be able to capture logs.