[Event Hubs Client] Idempotent Error Handling
The focus of these changes is to ensure that the AMQP connection and link
states are fully reset on any exception after retries are exhausted in order
to ensure no data loss.

In error scenarios with ambiguous outcomes, it is possible for the client and
service state for idempotent sequencing to become out-of-sync with no way to
detect or correct the problem.  When this happens, it is possible that events
being published are incorrectly identified by the service as duplicates and
their receipt acknowledged.  This leads the client to signal to callers that
publishing was successful, though the events were ignored by the service and
will never be available to be read.

By resetting the connection and forcing the client to use a new publisher
identifier, the service will reset state and perform the initial handshake
with the client, ensuring that state is properly synchronized between them.