[Event Hubs] Solution Infrastructure Updates
The focus of these changes is a reorganization of the Visual Studio solutions
for the current generation of Event Hubs packages, moving from individual
solutions to a single solution with individual filter sets.

Previously, each package had a dedicated solution to allow a focused view
when working, which required any external project dependencies to be updated
multiple times.  By moving to a unified solution, these items can be adjusted
a single time.