fix(resourcemanager): random return value of string.Compare
The docs of several `CompareTo` APIs says the return value is 0, 1, and -1. But actually for some use cases, the return value could be other numbers. The reason is that the one depending API `string.Compare` only ensure the return value will `Less than zero`, `Zero`, and `Greater than zero`. We've found under some environment, our test cases will fail because the return value is 10, -10, 8, etc.

The fix is to normalize the return value of `string.Compare` to 0, 1, and -1.